GraphImplementation filter added

+++ core/include/core/GraphImplementation/symmetric_matrix_graph.h
@@ -2,6 +2,8 @@
 #define GRAPH_CORE_MATRIX_GRAPH_H
 
 #include "core/graph_base.h"
+#include "core/macroses.h"
+
 namespace graphcpp
 {
 	class SymmetricMatrixGraph : public GraphBase
@@ -9,14 +11,14 @@ namespace graphcpp
 	private:
 		std::vector<std::vector<bool>> _matrix;
 	public:
-		SymmetricMatrixGraph(const std::vector<Edge>& edges, unsigned int dimension);
+		SymmetricMatrixGraph(const std::vector<Edge>& edges, ushort dimension);
 		//SymmetricMatrixGraph(const std::vector<std::vector<bool>>& matrix);
 
-		unsigned int Dimension() const noexcept override;
+		ushort Dimension() const noexcept override;
 
 		std::vector<std::vector<bool>> GetMatrix() const noexcept override;
 		std::vector<Edge> GetEdges() const noexcept override;
-		std::vector<int> GetLinkedVertexes(unsigned int vertex) const noexcept override;
+		std::vector<ushort> GetLinkedVertexes(ushort vertex) const noexcept override;
 	};
 }
 #endif
